add show booking schedule and add csrf.token for method post and fix bug for update customer (rewardPoints, password, customerType when updating)..

+++ src/main/java/com/lullabyhomestay/homestay_management/domain/dto/CustomerDTO.java
@@ -32,7 +32,7 @@ public class CustomerDTO {
 
     @NotBlank(message = "Vui lòng nhập địa chỉ")
     private String address;
-    private Float rewardPoints;
+    private Double rewardPoints;
     private Boolean isLocked;
     private CustomerType customerType;
 }
